Move GeneratePatchBytesForDelete to controller_ref_manager
